♻️ refactor: clean up m4 code
