--- downsample_rs/src/m4/mod.rs
@@ -1,5 +1,0 @@
-pub mod scalar;
-pub use scalar::*;
-pub mod simd;
-pub use simd::*;
-mod generic; // private module
